Move database context into separate namespace
