+++ Zilon.Core/CDT.LAST.MonoGameClient/Database/PlayerScore.cs
@@ -1,0 +1,4 @@
+﻿namespace CDT.LAST.MonoGameClient
+{
+    internal record PlayerScore(int Id, string NickName, uint RatingPosition, int Score);
+}
