+++ Zilon.Core/CDT.LAST.MonoGameClient/Program.cs
@@ -2,6 +2,7 @@
 using System.Globalization;
 using System.Threading;
 
+using CDT.LAST.MonoGameClient.Database;
 using CDT.LAST.MonoGameClient.Engine;
 using CDT.LAST.MonoGameClient.Screens;
 using CDT.LAST.MonoGameClient.ViewModels.MainScene;
