+++ Zilon.Core/CDT.LAST.MonoGameClient/Screens/LeaderBoardScreen.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
+using CDT.LAST.MonoGameClient.Database;
 using CDT.LAST.MonoGameClient.Engine;
 using CDT.LAST.MonoGameClient.Resources;
